pyproject.toml
[project]
name = "pipeline-bacen"
version = "0.1.0"
description = ""
authors = [
    {name = "VinicinCanedo",email = "vinicanedo@hotmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "requests (>=2.32.4,<3.0.0)",
    "streamlit (>=1.45.1,<2.0.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "supabase (>=2.15.3,<3.0.0)",
]

[tool.poetry]
packages = [{include = "pipeline_bacen", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
